refactor: components folders and exports
