+++ src/components/MultipleOptionResults/MultipleOptionResults.tsx
