+++ src/components/Quiz/Quiz.tsx
@@ -6,13 +6,12 @@ import QuizItemsTypes from "../../models/QuizItemsTypes.ts";
 import QuizItem from "../../models/QuizItem.ts";
 import IQuizIterator from "../../models/IQuizIterator.ts";
 import MultipleOptionQuestion from "../../models/MultipleOptionQuestion.ts";
-import PresentationSlide from "../../models/PresentationSlide.ts";
 import handlers from "../../handlers";
-import SingleAnswer from "../User/SingleAnswer.tsx";
-import MultipleAnswer from "../User/MultipleAnswer.tsx";
-import Presentation from "../Presentation/Presentation.tsx";
-import MultipleOptionResults from "../Admin/MultipleOptionResults.tsx";
-import QuizStart from "../User/QuizStart.tsx";
+import SingleAnswer from "../SingleAnswer";
+import MultipleAnswer from "../MultipleAnswer";
+import Presentation from "../Presentation";
+import MultipleOptionResults from "../MultipleOptionResults";
+import QuizStart from "../QuizStart";
 
 const Quiz = () => {
     const [quiz, setQuiz] = useState(
@@ -77,7 +76,7 @@ const Quiz = () => {
             break;
         case QuizItemsTypes.Presentation:
             BodyComponent = (
-                <Presentation presentationItem={question.item as PresentationSlide} />
+                <Presentation />
             );
             break;
         case QuizItemsTypes.SingleAnswer:
